Add new NotificationService
This is an IntentService for handling notifications more
_intelligently_; still has few bugs that needs to be fixed.

+++ app/src/main/java/islam/adhanalarm/Preferences.java
@@ -65,8 +65,8 @@ public class Preferences {
         editor.apply();
     }
 
-    public int getNotificationMethod(short time) {
-        return mSharedPreferences.getInt(KEY_NOTIFICATION_METHOD.concat(Short.toString(time)),
+    public int getNotificationMethod(int time) {
+        return mSharedPreferences.getInt(KEY_NOTIFICATION_METHOD.concat(Integer.toString(time)),
                 CONSTANT.SUNRISE == time ? CONSTANT.NOTIFICATION_NONE
                         : CONSTANT.NOTIFICATION_DEFAULT);
     }
